requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
websockets>=12.0
numpy>=1.24.0
python-dotenv>=1.0.0
openai>=1.0.0
httpx>=0.25.0
huggingface-hub>=0.20.0
matplotlib>=3.7.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
faster-whisper>=1.1.0
onnxruntime>=1.16.0
mcp>=1.0.0
croniter>=2.0.0
# audioop-lts>=0.2.1  # Only needed for Python 3.13+ (audioop is built-in for earlier versions)
